Basic GUI in OpenGL
So, I setup a basic GUI using OpenGL with a nxn grid coordinate system
with mouse-click interaction and a simulated command line interface
(i.e. type commands into window and enter to send commands).

To run, you must have GLUT toolkit installed to manipulate OpenGL
libraries. Only tested on Visual Studio 2012. Should work on UNIX/LINUX
systems and MacOS systems configured with GLUT.

This was more to teach myself OpenGl than actually design something for
the project but it could be a basis to work off of if we do decided to
use OpenGL.

I could not, for the life of me, get GLUT toolkit installed on the
UNIX/LINUX servers, but did manage it on Windows.

Due to these reasons, and since we have not yet decided which graphical
packages we plan to use, I stuck this file in the supplementary
documents folder.