Added basic if-else "parser" and helper functions
Added a simple parser to test functions. Having issues with AI and
altering color and difficulty (keep getting errors). Also we need to
figure out a way to compare valid moves (not legal) when it is sent to
parser. I have a method but is messy, the file is included so take a
look