Basic GUI in OpenGL now implemented into client.
Alright. So it's messy as fuck. I know. Sorry. I could not, for the life of me, encapsulate the OpenGL GUI stuff into a class and ended up just shoving it onto the backend of the game_client.cpp.

Makefile should be set and good for compile. Simply make and the run main.